+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- instant scrolling to top on active `PanelSideBarItem`; scrolling smoothing to center instead.
+
 ## [5.3.0] - 2025-09-30
 
 ### Fix